Updated bookstore in server

+++ server/1ªEvaluacion/bookstore/src/main/java/fpmislata/bookstore/b_presentation/admin/model/AuthorCollection.java
@@ -1,4 +1,4 @@
 package fpmislata.bookstore.b_presentation.admin.model;
 
-public record AuthorCollection(Integer id, String name) {
+public record AuthorCollection(Long id, String name) {
 }
